[FEAT] 이벤트, 공지사항 feature 통일 | Merge event, notice feature TALEARNT-491
