+++ src/features/eventNotice/eventNotice.store.ts
@@ -6,3 +6,8 @@ export const useEventPageStore = create<pageStoreType>(set => ({
   page: 1,
   setPage: page => set({ page }),
 }));
+
+export const useNoticePageStore = create<pageStoreType>(set => ({
+  page: 1,
+  setPage: page => set({ page }),
+}));
