--- src/features/notice/notice.store.ts
@@ -1,8 +1,0 @@
-import { create } from "zustand";
-
-import { pageStoreType } from "@shared/type/api.type";
-
-export const useNoticePageStore = create<pageStoreType>(set => ({
-  page: 1,
-  setPage: page => set({ page }),
-}));
