--- src/features/notice/notice.type.ts
@@ -1,7 +1,0 @@
-export type noticeType = {
-  noticeNo: number;
-  title: string;
-  noticeType: string;
-  content: string;
-  createdAt: string;
-};
